requirements.txt
flask
requests
openai
PyGithub
python-dotenv